+++ .github/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+-   Add batch accept script
+
 ## [2.4.0](https://github.com/hackmcgill/hackerapi/tree/2.4.0) - 2020-01-31
 
 ### Added
@@ -12,9 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added the live site link to the check in email
 
 ## [2.3.1](https://github.com/hackmcgill/hackerapi/tree/2.3.1) - 2020-01-15
-
-### Added
--   Added the live site link to the check in email
 
 ### Removed